WebSocket KeepAlive Refactoring

+++ src/test/java/trade/wayruha/mexc/ws/PrivateWSConnectionTest.java
@@ -50,7 +50,7 @@ public class PrivateWSConnectionTest {
     }
 
     @SneakyThrows
-    private static void sleep(int timeMs){
+    private static void sleep(long timeMs){
         Thread.sleep(timeMs);
     }
